+++ Tests/MongoSwiftTests/AsyncTestUtils.swift
@@ -1,0 +1,19 @@
+import Foundation
+import MongoSwift
+import TestsCommon
+
+extension MongoClient {
+    static func makeTestClient(
+        _ uri: String = MongoSwiftTestCase.connStr,
+        options: ClientOptions? = nil
+    ) throws -> MongoClient {
+        var opts = options ?? ClientOptions()
+        if MongoSwiftTestCase.ssl {
+            opts.tlsOptions = TLSOptions(
+                caFile: URL(string: MongoSwiftTestCase.sslCAFilePath ?? ""),
+                pemFile: URL(string: MongoSwiftTestCase.sslPEMKeyFilePath ?? "")
+            )
+        }
+        return try MongoClient(uri, options: opts)
+    }
+}
